Moving some code from protocol to engine/server
