+++ Unix/server/servercommon.c
@@ -14,10 +14,7 @@ static Options *s_optsPtr = NULL;
 static ServerData *s_dataPtr = NULL;
 static ServerType serverType = OMI_SERVER;
 static const char* arg0 = 0;
-
-#if defined(_ENGINE)
-static Lock s_disp_mutex = LOCK_INITIALIZER;
-#endif
+static Lock *s_disp_mutex_ptr = NULL;
 
 void PrintProviderMsg(_In_ Message* msg)
 {
@@ -59,30 +56,6 @@ void PrintProviderMsg(_In_ Message* msg)
     }
 #endif // !defined(CONFIG_FAVORSIZE)
 }
-
-/*
-    Simple interaction object to respond to the noop request.
-    It just sends a noop response and closes the interaction
-    (therefore shutting down)
-*/
-void _NoopInteractionAck( _In_ Strand* self) 
-{
-    // do nothing
-}
-
-StrandFT _NoopInteractionUserFT = { 
-        NULL, 
-        NULL, 
-        _NoopInteractionAck, 
-        NULL, 
-        NULL, 
-        NULL,
-        NULL,
-        NULL, 
-        NULL, 
-        NULL, 
-        NULL, 
-        NULL };
 
 void GetCommandLineDestDirOption(
     int* argc_,
@@ -838,12 +811,13 @@ void GetConfigFileOptions()
     return;
 }
 
-void SetDefaults(Options *opts_ptr, ServerData *data_ptr, const char *executable, ServerType type)
+void SetDefaults(Options *opts_ptr, ServerData *data_ptr, const char *executable, ServerType type, Lock *disp_mutex_ptr)
 {
     s_optsPtr = opts_ptr;
     s_dataPtr = data_ptr;
     arg0 = executable;
     serverType = type;
+    s_disp_mutex_ptr = disp_mutex_ptr;
 
     memset(s_dataPtr, 0, sizeof(ServerData));
 
@@ -885,100 +859,7 @@ void SetDefaults(Options *opts_ptr, ServerData *data_ptr, const char *executable
     s_dataPtr->internalSock = INVALID_SOCK;
 }
 
-#if defined(_ENGINE)
-STRAND_DEBUGNAME( NoopRequest )
-
-static void _ProcessNoopRequest(_Inout_ InteractionOpenParams*  params)
-{
-    Strand* strand;
-    NoOpReq* req = (NoOpReq*)params->msg;
-    NoOpRsp* rsp;
-
-#if !defined(CONFIG_FAVORSIZE)
-    if (s_optsPtr->trace)
-    {
-        NoOpReq_Print(req, stdout);
-    }
-#endif
-            
-    strand = Strand_New( STRAND_DEBUG( NoopRequest ) &_NoopInteractionUserFT, 0, STRAND_FLAG_ENTERSTRAND, params );
-
-    if( NULL == strand )
-    {
-        trace_OutOfMemory();
-        Strand_FailOpen(params);
-        err(ZT("out of memory"));
-    }
-
-    /* Send NoOp response back */
-    rsp = NoOpRsp_New(req->base.base.operationId);
-
-    if (!rsp)
-    {
-        trace_OutOfMemory();
-        Strand_FailOpen(params);
-        err(ZT("out of memory"));
-    }
-
-#if !defined(CONFIG_FAVORSIZE)
-    if (s_optsPtr->trace)
-    {
-        NoOpRsp_Print(rsp, stdout);
-    }
-#endif
-
-    Strand_Ack( strand );   // Ack open msg
-    Strand_Post( strand, &rsp->base );
-    Strand_Close( strand );   
-    Strand_Leave( strand);
-    
-    NoOpRsp_Release(rsp);
-
-    trace_ServerReceivedNoOpReqTag( (int)s_optsPtr->terminateByNoop );
-    
-    if (s_optsPtr->terminateByNoop)
-    {
-        s_dataPtr->terminated = MI_TRUE;
-        Selector_StopRunning(&s_dataPtr->selector);
-    }
-}
-
 /* Called by protocol stack to dispatch an incoming request message */
-static void _RequestCallback(
-    _Inout_ InteractionOpenParams* interactionParams)
-{
-    ServerCallbackData* self = (ServerCallbackData*)interactionParams->callbackData;
-    Message* msg = interactionParams->msg;
-    MI_Result result = MI_RESULT_OK;
-
-    DEBUG_ASSERT( NULL != interactionParams );
-    DEBUG_ASSERT( NULL != msg );
-    
-    if (NoOpReqTag == msg->tag)
-    {
-        _ProcessNoopRequest(interactionParams);
-        return;
-    }
-
-#if !defined(CONFIG_FAVORSIZE)
-    if (s_optsPtr->trace)
-    {
-        MessagePrint(msg, stdout);
-    }
-#endif
-    
-    Lock_Acquire(&s_disp_mutex);
-    result = Disp_HandleInteractionRequest(
-                &self->data->disp, 
-                interactionParams );
-    Lock_Release(&s_disp_mutex);
-    if( result != MI_RESULT_OK )
-    {
-        Strand_FailOpenWithResult(interactionParams, result, PostResultMsg_NewAndSerialize);
-    }
-}
-#endif /*_ENGINE */
- 
 void FUNCTION_NEVER_RETURNS err(const ZChar* fmt, ...)
 {
     va_list ap;
@@ -1060,87 +941,6 @@ MI_Result InitializeNetwork()
     return r;
 }
 
-#if defined(_ENGINE)
-MI_Result WsmanProtocolListen()
-{
-    MI_Result r = MI_RESULT_OK;
-
-    /* Set WSMAN options and create WSMAN server */
-    s_dataPtr->wsman_size = s_optsPtr->httpport_size + s_optsPtr->httpsport_size;
-    if ( s_dataPtr->wsman_size > 0 )
-    {
-        s_dataPtr->wsman = PAL_Calloc(s_dataPtr->wsman_size, sizeof(WSMAN*));
-        if ( s_dataPtr->wsman == NULL )
-        {
-            err(ZT("memory allocation failure allocating %d bytes"), s_dataPtr->wsman_size * sizeof(WSMAN*));
-        }
-    }
-
-    {
-        int wsman_count = 0;
-        WSMAN_Options options = DEFAULT_WSMAN_OPTIONS;
-#if !defined(CONFIG_FAVORSIZE)
-        options.enableTracing = s_optsPtr->trace;
-#endif
-        options.enableHTTPTracing = s_optsPtr->httptrace;
-
-        /* Start up the non-encrypted listeners */
-        int count;
-        for ( count = 0; count < s_optsPtr->httpport_size; ++count )
-        {
-            r = WSMAN_New_Listener(
-                &s_dataPtr->wsman[wsman_count++],
-                &s_dataPtr->selector, 
-                s_optsPtr->httpport[count],
-                0,
-                s_optsPtr->sslCipherSuite,
-                s_optsPtr->sslOptions,
-                _RequestCallback,
-                &s_dataPtr->wsmanData,
-                &options);
-
-            if (r != MI_RESULT_OK)
-            {
-                err(ZT("WSMAN_New_Listener() failed for port %u. Err = %d"), s_optsPtr->httpport[count], r);
-            }
-
-            /* Log start up message */
-            trace_ListeningOnPort(s_optsPtr->httpport[count]);
-        }
-
-        /* Start up the encrypted listeners */
-        for ( count = 0; count < s_optsPtr->httpsport_size; ++count )
-        {
-            r = WSMAN_New_Listener(
-                &s_dataPtr->wsman[wsman_count++],
-                &s_dataPtr->selector, 
-                0,
-                s_optsPtr->httpsport[count],
-                s_optsPtr->sslCipherSuite,
-                s_optsPtr->sslOptions,
-                _RequestCallback,
-                &s_dataPtr->wsmanData,
-                &options);
-
-            if (r != MI_RESULT_OK)
-            {
-                err(ZT("WSMAN_New_Listener() failed for encrypted port %u. Err = %d"), s_optsPtr->httpsport[count], r);
-            }
-
-            /* Log start up message */
-            trace_ListeningOnEncryptedPort(s_optsPtr->httpsport[count]);
-        }
-    }
-
-    return r;
-}
-#else
-static void _DummyCallback( _Inout_ InteractionOpenParams *params)
-{
-    err(ZT("_DummyCallback reached"));
-}
-#endif
-
 MI_Result BinaryProtocolListenFile(
     const char *socketFile, 
     MuxIn *mux, 
@@ -1149,15 +949,17 @@ MI_Result BinaryProtocolListenFile(
 {
     MI_Result r = MI_RESULT_OK;
 
-#if defined(_ENGINE)
     /* mux */
     {
-        if(MuxIn_Init(mux, _RequestCallback, &s_dataPtr->protocolData, NULL, PostResultMsg_NewAndSerialize) != MI_RESULT_OK)
+#if defined(_ENGINE)
+        if(MuxIn_Init(mux, EngineCallback, &s_dataPtr->protocolData, NULL, PostResultMsg_NewAndSerialize) != MI_RESULT_OK)
+#else
+        if(MuxIn_Init(mux, ServerCallback, &s_dataPtr->protocolData, NULL, NULL) != MI_RESULT_OK)
+#endif
         {
             err(ZT("MuxIn_Init() failed"));
         }
     }
-#endif
         
     /* Create new protocol object */
     {
@@ -1165,13 +967,9 @@ MI_Result BinaryProtocolListenFile(
             protocol, 
             &s_dataPtr->selector, 
             socketFile,
-#if defined(_ENGINE)
             MuxIn_Open,
             mux);
-#else
-            _DummyCallback,
-            NULL);
-#endif
+
         if (r != MI_RESULT_OK)
         {
             err(ZT("Protocol_New_Listener() failed: %T"), socketFile);
@@ -1200,15 +998,17 @@ MI_Result BinaryProtocolListenSock(
 {
     MI_Result r = MI_RESULT_OK;
 
-#if defined(_ENGINE)
     /* mux */
     {
-        if(MuxIn_Init(mux, _RequestCallback, &s_dataPtr->protocolData, NULL, PostResultMsg_NewAndSerialize) != MI_RESULT_OK)
+#if defined(_ENGINE)
+        if(MuxIn_Init(mux, EngineCallback, &s_dataPtr->protocolData, NULL, PostResultMsg_NewAndSerialize) != MI_RESULT_OK)
+#else
+        if(MuxIn_Init(mux, ServerCallback, &s_dataPtr->protocolData, NULL, NULL) != MI_RESULT_OK)
+#endif
         {
             err(ZT("MuxIn_Init() failed"));
         }
     }
-#endif
         
     /* Create new protocol object */
     {
@@ -1216,13 +1016,8 @@ MI_Result BinaryProtocolListenSock(
             protocol, 
             &s_dataPtr->selector, 
             sock,
-#if defined(_ENGINE)
             MuxIn_Open,
             mux);
-#else
-            _DummyCallback,
-            NULL);
-#endif
 
         if (r != MI_RESULT_OK)
         {
@@ -1270,10 +1065,10 @@ MI_Result RunProtocol()
         if (s_dataPtr->reloadDispFlag || 
             reload_file_exists == 0)
         {
-            Lock_Acquire(&s_disp_mutex);
+            Lock_Acquire(s_disp_mutex_ptr);
             Disp_Reload(&s_dataPtr->disp);
             s_dataPtr->reloadDispFlag = MI_FALSE;
-            Lock_Release(&s_disp_mutex);
+            Lock_Release(s_disp_mutex_ptr);
 
             if (reload_file_exists == 0)
             {
